apps/apps_requirements.txt
discord
craiyon.py
g4f
# g4f==0.2.4.1

pynput
Pillow
pytesseract
pyperclip
imageio
sounddevice
pydub